+++ alacritty/src/config/color.rs
@@ -18,16 +18,17 @@ pub struct Colors {
     pub line_indicator: LineIndicatorColors,
     pub hints: HintColors,
     pub transparent_background_colors: bool,
+    pub draw_bold_text_with_bright_colors: bool,
     footer_bar: BarColors,
 }
 
 impl Colors {
     pub fn footer_bar_foreground(&self) -> Rgb {
-        self.search.bar.foreground.or(self.footer_bar.foreground).unwrap_or(self.primary.background)
+        self.footer_bar.foreground.unwrap_or(self.primary.background)
     }
 
     pub fn footer_bar_background(&self) -> Rgb {
-        self.search.bar.background.or(self.footer_bar.background).unwrap_or(self.primary.foreground)
+        self.footer_bar.background.unwrap_or(self.primary.foreground)
     }
 }
 
@@ -126,8 +127,6 @@ impl Default for InvertedCellColors {
 pub struct SearchColors {
     pub focused_match: FocusedMatchColors,
     pub matches: MatchColors,
-    #[config(deprecated = "use `colors.footer_bar` instead")]
-    bar: BarColors,
 }
 
 #[derive(ConfigDeserialize, Debug, Copy, Clone, PartialEq, Eq)]