+++ alacritty_terminal/src/ansi.rs
@@ -52,11 +52,7 @@ impl<'de> serde::Deserialize<'de> for CursorShapeShim {
 }
 
 impl alacritty_config::SerdeReplace for CursorShapeShim {
-    fn replace(
-        &mut self,
-        key: &str,
-        value: serde_yaml::Value,
-    ) -> Result<(), Box<dyn std::error::Error>> {
+    fn replace(&mut self, key: &str, value: toml::Value) -> Result<(), Box<dyn std::error::Error>> {
         if !key.is_empty() {
             return Err(format!("Fields \"{0}\" do not exist", key).into());
         }